Extract into files and refactor
