+++ package.json
@@ -7,7 +7,9 @@
     "node": ">=14.0.0"
   },
   "scripts": {
-    "start": "node src/index.js"
+    "start": "node src/index.js",
+    "discord": "node src/index.js discord",
+    "twitch": "node src/index.js twitch"
   },
   "keywords": [],
   "author": "",
